Fix circular dependency between analytics and service connection
